+++ apps/server/src/routes/asignaciones.ts
@@ -1,19 +1,4 @@
-// tryCatch helper
-type Success<T> = { data: T; error: null };
-type Failure<E> = { data: null; error: E };
-type Result<T, E = Error> = Success<T> | Failure<E>;
-
-export async function tryCatch<T, E = Error>(
-  promise: Promise<T>,
-): Promise<Result<T, E>> {
-  try {
-    const data = await promise;
-    return { data, error: null };
-  } catch (error) {
-    return { data: null, error: error as E };
-  }
-}
-
+import { tryCatch } from "../trycatch";
 import { db } from "../db/server";
 import { asignaciones } from "../db/schema";
 import { z } from "zod";
